Fix for the gf shell. Supports newer versions of IPython, v3.11+ (#414)
IPython.terminal broke the greatfet shell such that any attempt to use it resulted in an exception ending in:
Exception 'NoneType' object has no attribute 'check_complete'

This was traced to a change in IPython, version 8.11 and has broken other projects as well, refer to this issue in the IPython project: ipython/ipython#13966

This PR fixes the problem, as suggested in the issue report (linked above). It should be compatible with IPython v8.11 and later, and also backwards compatible for earlier versions of the library as well (as per the above issues last comment).

I tested the raw Python script (greatefet_shell.py) but have not managed to generate an install (.egg ?) using the README instructions. The .py file, when run directly, no longer creates an exception and the shell runs normally.