Fix str vs bytes issue in Console
Fixes a bug in greatfet console that causes runtime errors on Windows and in certain cases on nix-based OSes when using greatfet uart functionality.

NTConsole (and in certain cases POSIXConsole) emit `str` types when keys are pressed, `greatfet_uart.py` is written to expect keys as `bytestring`. Since Console is currently only used by the UART functionality, this commit changes the Console to return only `bytestring`s when polled for keypresses.